Correções na comunicação, na biblioteca do imu e no gimbal

+++ src/Communication.h
@@ -31,7 +31,7 @@ class Communication {
   void writeBytes(uint8_t* data, uint8_t size);
   uint8_t checksum(uint8_t cmd, uint8_t size, uint8_t* data, uint8_t dataSize);
 
-  void updateGoals(int16_t* pos, uint8_t size);
+  void updateGoals(uint8_t* pos, uint8_t size);
 
   static const uint8_t serialBufferSize = 50;
